feat: change amount field type from int to float in create and update ration input DTOs

+++ Antonieta.Backend/dtos/create_ration_input_dto.py
@@ -6,6 +6,6 @@ class create_ration_input_dto(BaseModel):
     Data Transfer Object (DTO) para criar registro de entrada de ração.
     """
     ration_stock_id: int
-    amount: int
+    amount: float
     date: datetime
     description: str
